Remove HitReceptor from TouchNote, and handle input directly
